:rocket: Add commands structure with clean and reset command
