+++ src/commands/help.js
@@ -1,0 +1,5 @@
+const helpCommand = () => {
+  console.log('Help command coming soon.');
+};
+
+export default helpCommand;
