+++ src/config/settings.js
@@ -1,0 +1,4 @@
+import path from 'path';
+import os from 'os';
+
+export const externalConfigLocation = path.resolve(`${os.homedir()}/.fancy.json`);
